fix(Step5) : 프로그램 play 버그 수정
프로그램 실행 하면서 전부 움직이지 않는 버그를 발견하여 수정함 : 
  - 주입한 randomEngine... threshold값이 10 이었음 (무조건 전진 못함...)
  - threshold값 수정
  - 클래스 길어진거 보기 싫어서 RandomEngine 추출함.
  - 전체적으로 import 정리 및 라인 정리
