+++ src/main/java/racingGame/RacingGamePlayer.java
@@ -7,29 +7,10 @@ import racingGame.ui.InputView;
 import racingGame.ui.ResultView;
 
 import java.util.Arrays;
-import java.util.Random;
 import java.util.stream.Collectors;
 
 public class RacingGamePlayer {
-	private static final Engine randomEngine = new Engine() {
-		private static final int RANDOM_RANGE = 10;
-		private static final int GO_THRESHOLD = 10;
-
-		private Random random = new Random();
-
-		@Override
-		public int moveDistanceForOneTime() {
-			if (canGo()) {
-				return 1;
-			}
-
-			return 0;
-		}
-
-		private boolean canGo() {
-			return random.nextInt(RANDOM_RANGE) > GO_THRESHOLD;
-		}
-	};
+	private static final Engine randomEngine = new RandomEngine();
 
 	private InputView inputView;
 	private ResultView resultView;
