+++ src/main/java/racingGame/domain/RacingGame.java
@@ -2,7 +2,6 @@ package racingGame.domain;
 
 import java.util.Arrays;
 import java.util.List;
-import java.util.Map;
 import java.util.stream.Collectors;
 
 public class RacingGame {
